requirements.txt
streamlit
streamlit-pdf-viewer
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-ollama 
ollama
langchain-core
faiss-cpu
sentence-transformers
huggingface-hub
pymupdf
torch
transformers
pillow
accelerate




